+++ src/engine.rs
@@ -228,6 +228,47 @@ where
         Ok(None)
     }
 
+    pub fn scan_messages<S, C>(
+        &self,
+        region_id: u64,
+        start_key: Option<&[u8]>,
+        end_key: Option<&[u8]>,
+        reverse: bool,
+        callback: C,
+    ) -> Result<()>
+    where
+        S: Message,
+        C: FnMut(&[u8], S) -> bool,
+    {
+        let _t = StopWatch::new(&*ENGINE_READ_MESSAGE_DURATION_HISTOGRAM);
+        if let Some(memtable) = self.memtables.get(region_id) {
+            memtable
+                .read()
+                .scan_messages(start_key, end_key, reverse, callback)?;
+        }
+        Ok(())
+    }
+
+    pub fn scan_raw_messages<C>(
+        &self,
+        region_id: u64,
+        start_key: Option<&[u8]>,
+        end_key: Option<&[u8]>,
+        reverse: bool,
+        callback: C,
+    ) -> Result<()>
+    where
+        C: FnMut(&[u8], &[u8]) -> bool,
+    {
+        let _t = StopWatch::new(&*ENGINE_READ_MESSAGE_DURATION_HISTOGRAM);
+        if let Some(memtable) = self.memtables.get(region_id) {
+            memtable
+                .read()
+                .scan_raw_messages(start_key, end_key, reverse, callback)?;
+        }
+        Ok(())
+    }
+
     pub fn get_entry<M: MessageExt>(
         &self,
         region_id: u64,
@@ -610,7 +651,13 @@ mod tests {
             RaftLogEngine::open_with(cfg, file_system, listeners).unwrap()
         }
 
-        fn scan<FR: Fn(u64, LogQueue, &[u8])>(&self, rid: u64, start: u64, end: u64, reader: FR) {
+        fn scan_entries<FR: Fn(u64, LogQueue, &[u8])>(
+            &self,
+            rid: u64,
+            start: u64,
+            end: u64,
+            reader: FR,
+        ) {
             let mut entries = Vec::new();
             self.fetch_entries_to::<Entry>(
                 rid,
@@ -695,7 +742,7 @@ mod tests {
             for i in 10..20 {
                 let rid = i;
                 let index = i;
-                engine.scan(rid, index, index + 2, |_, q, d| {
+                engine.scan_entries(rid, index, index + 2, |_, q, d| {
                     assert_eq!(q, LogQueue::Append);
                     assert_eq!(d, &data);
                 });
@@ -706,7 +753,7 @@ mod tests {
             for i in 10..20 {
                 let rid = i;
                 let index = i;
-                engine.scan(rid, index, index + 2, |_, q, d| {
+                engine.scan_entries(rid, index, index + 2, |_, q, d| {
                     assert_eq!(q, LogQueue::Append);
                     assert_eq!(d, &data);
                 });
@@ -756,7 +803,7 @@ mod tests {
 
                     let engine = engine.reopen();
                     if let Some((start, end)) = *steps.last().unwrap() {
-                        engine.scan(rid, start, end, |_, _, d| {
+                        engine.scan_entries(rid, start, end, |_, _, d| {
                             assert_eq!(d, &data);
                         });
                     } else {
@@ -766,7 +813,7 @@ mod tests {
                     engine.purge_manager.must_rewrite_append_queue(None, None);
                     let engine = engine.reopen();
                     if let Some((start, end)) = *steps.last().unwrap() {
-                        engine.scan(rid, start, end, |_, _, d| {
+                        engine.scan_entries(rid, start, end, |_, _, d| {
                             assert_eq!(d, &data);
                         });
                     } else {
@@ -870,6 +917,14 @@ mod tests {
         engine.write(&mut batch_2.clone(), true).unwrap();
         let engine = engine.reopen();
         assert_eq!(engine.get(rid, &key).unwrap(), v2);
+        let mut res = vec![];
+        engine
+            .scan_raw_messages(rid, Some(&key), None, false, |key, value| {
+                res.push((key.to_vec(), value.to_vec()));
+                true
+            })
+            .unwrap();
+        assert_eq!(res, vec![(key.clone(), v2.clone())]);
 
         // put | delete | put |
         //                    ^ rewrite
@@ -920,7 +975,7 @@ mod tests {
         compact_log.add_command(rid, Command::Compact { index: 5 });
         engine.write(&mut compact_log, true).unwrap();
         let engine = engine.reopen();
-        engine.scan(rid, 5, 10, |_, q, d| {
+        engine.scan_entries(rid, 5, 10, |_, q, d| {
             assert_eq!(q, LogQueue::Append);
             assert_eq!(d, &data);
         });
@@ -944,7 +999,7 @@ mod tests {
         engine.memtables.apply_append_writes(compact_log.drain());
         engine.purge_manager.must_rewrite_rewrite_queue();
         let engine = engine.reopen();
-        engine.scan(rid, 10, 25, |_, q, d| {
+        engine.scan_entries(rid, 10, 25, |_, q, d| {
             assert_eq!(q, LogQueue::Append);
             assert_eq!(d, &data);
         });
@@ -957,7 +1012,7 @@ mod tests {
             .purge_manager
             .must_rewrite_append_queue(None, Some(2));
         let engine = engine.reopen();
-        engine.scan(rid, 10, 25, |_, q, d| {
+        engine.scan_entries(rid, 10, 25, |_, q, d| {
             assert_eq!(q, LogQueue::Append);
             assert_eq!(d, &data);
         });
@@ -978,7 +1033,7 @@ mod tests {
         compact_log.add_command(rid, Command::Compact { index: 20 });
         engine.write(&mut compact_log, true).unwrap();
         let engine = engine.reopen();
-        engine.scan(rid, 20, 25, |_, q, d| {
+        engine.scan_entries(rid, 20, 25, |_, q, d| {
             assert_eq!(q, LogQueue::Append);
             assert_eq!(d, &data);
         });
@@ -997,7 +1052,7 @@ mod tests {
             .purge_manager
             .must_rewrite_append_queue(None, Some(2));
         let engine = engine.reopen();
-        engine.scan(rid, 10, 15, |_, q, d| {
+        engine.scan_entries(rid, 10, 15, |_, q, d| {
             assert_eq!(q, LogQueue::Append);
             assert_eq!(d, &data);
         });
@@ -1152,7 +1207,7 @@ mod tests {
 
         // All entries should be available.
         for rid in 1..=10 {
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &data);
             });
         }
@@ -1164,7 +1219,7 @@ mod tests {
         assert_eq!(engine.memtables.cleaned_region_ids(), cleaned_region_ids);
 
         for rid in 1..=10 {
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &data);
             });
         }
@@ -1268,7 +1323,7 @@ mod tests {
 
         let engine = engine.reopen();
         for rid in 1..21 {
-            engine.scan(rid, 1, 21, |_, _, d| {
+            engine.scan_entries(rid, 1, 21, |_, _, d| {
                 assert_eq!(d, &data);
             });
         }
@@ -1303,7 +1358,7 @@ mod tests {
 
         let engine = engine.reopen();
         for rid in 1..=3 {
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &data);
             });
         }
@@ -1526,12 +1581,12 @@ mod tests {
 
         let engine = RaftLogEngine::open_with_file_system(cfg, fs).unwrap();
         for rid in 1..25 {
-            engine.scan(rid, 1, 6, |_, _, d| {
+            engine.scan_entries(rid, 1, 6, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }
         for rid in 25..=50 {
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }
@@ -1591,7 +1646,7 @@ mod tests {
             if existing_emptied.contains(&rid) || incoming_emptied.contains(&rid) {
                 continue;
             }
-            engine.scan(rid, 1, 6, |_, _, d| {
+            engine.scan_entries(rid, 1, 6, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }
@@ -1599,14 +1654,14 @@ mod tests {
             if existing_emptied.contains(&rid) || incoming_emptied.contains(&rid) {
                 continue;
             }
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }
         for rid in existing_emptied {
             let first_index = if rid < 25 { 1 } else { 6 };
             let last_index = if rid < 25 { 5 } else { 10 };
-            engine.scan(rid, first_index, last_index + 1, |_, _, d| {
+            engine.scan_entries(rid, first_index, last_index + 1, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }
@@ -1685,7 +1740,7 @@ mod tests {
 
         let engine = RaftLogEngine::open_with_file_system(cfg, fs).unwrap();
         for rid in 1..10 {
-            engine.scan(rid, 1, 11, |_, _, d| {
+            engine.scan_entries(rid, 1, 11, |_, _, d| {
                 assert_eq!(d, &entry_data);
             });
         }